pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecom-image-agent"
version = "0.2.0"
description = "FastAPI + React e-commerce image generation workspace."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "streamlit>=1.44.0",
  "langgraph>=0.3.18",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "pillow>=10.4.0",
  "PyYAML>=6.0.2",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "google-genai>=1.66.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "python-multipart>=0.0.9",
  "SQLAlchemy>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "PyJWT>=2.9.0",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.28.0",
  "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
